requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
pytz==2020.1
sqlparse==0.3.1
django-extensions==3.0.9
djangorestframework==3.11.1
Pillow==7.2.0
pytz==2020.1
sorl-thumbnail==12.6.3
sqlparse==0.3.1
suds-py3==1.4.1.0
